r344: Many changes in this changeset, but none very major.
- made the glosswords table "viewable" , e.g. .../rootcanal/rootcanal.pl/edit/glosswords
- added 'semkey' field to lexicon, glosswords, and etyma tables. (see sql/semkey.sql)
- added 'semkey' as a search/display field in lexicon, etyma, and glosswords "views"
- the 'status' field in the lexicon is now exposed, in the event we want to use it for keeping track of individual records needing help.
- the very slow 'updateprojects' runmode got even slower, but now does a better job of counting tags.
- "projects" have been reworked a bit, to try to bring that functionality into line with our current process (i.e. where "approving" is not a distinct workflow step)
- admin.tt is now two templates, admin.tt, and admin_div.tt, to facilitate using that code within other pages. A side effect, due to changing the targets of the <a> elements, is that most links will now open a new browser window.

Other changes:

working on a "multiwindow" display of etyma and lexicon records (a bit like /gnis, but for use as a tagging interface). it is currently a runmode named "widget", and it's not working very well at the moment.

there is now a "chapter_tweaker", a version of the "chapter_browser", for use with widget)

several other minor changes to core functionality:

there is now a new "search type" (in addition to "word", "int", "beginword", etc.). it is "value", and it is simply the value passed in "un-regexed", surrounded by quotes.