r345: the code which creates the morphemes table.
Basically, it works as follows:

mkm.sh is a sequence of calls to mysql and perl to extract files, massage them, and load bits back into the database.

The steps are as follows:

1. Extract the bits from the lexicon table that will be needed.
2. Run morphemes.pl, which runs syllablification station and creates a row for each morpheme found.
3. Run transduce2.pl, which tries to parse syllables into Init + Rhyme + Tone.
4. Create the morphemes table and load the whole mess into it.

Note that the shell script does not presume a username and pwd for mysql.

Takes about 4 mins to run on my MacBook.