r363: more "progress" details.
The detail report shows accepts over time (i.e. by year and month)

It is accessed via the Progress Report.

I put it on a separate page as I am thinking of generating some cute graphics using Google Graphs and did not want the
progress page to stop working merely because Google might not be available on the server from which the Reports are generated.