r554: Initial checkin of a experimental database interface implementing some UI conveniences, including but not limited to:
- autosuggest
- tabbed panes (for persistence)

Not everything works yet, and the autosuggest javascript used has a
few idiosyncrasies (features) which might even be called bugs.

From a development point of view, the implementation is pretty simple:

- only 23 lines of user-written javascript
- a single css file, 75 lines
- 7 php modules, totaling 262 lines
- uses jQuery for tabs, and the autosuggest is also jQuery-based

To deploy:

I think all one needs to do is ensure that PHP is installed and put all the files in
a webserver accessible directory (making sure they are executable).