r705: changes to support "autosequencing" of unsequenced etyma.
Not perfect! etyma are sequenced from 1000 to N, and are not sequence within their semkey groupings. (think
about it: what is the domain of the sequencing: is it V? V+F? V+F+C? The seqencing tool uses "whatever is
the value of the chapter of the record the user clicked on").

to resequence etyma (after things have been edit a bunch), visit the Tools page.

NB: there are db schema changes required by this enhancement: the etyma table needs refcount and seqlocked fields
added.